feat: get navigation menus directly from each layout for client and admin
